Refactor Core tests
