--- core/src/test/kotlin/com/hexagonkt/core/HexagonCoreSamplesTest.kt
@@ -1,53 +1,0 @@
-package com.hexagonkt.core
-
-import com.hexagonkt.core.logging.logger
-import com.hexagonkt.core.logging.Logger
-import com.hexagonkt.core.logging.LoggingLevel.*
-import com.hexagonkt.core.logging.LoggingManager
-import org.junit.jupiter.api.Test
-import kotlin.test.assertContains
-
-internal class HexagonCoreSamplesTest {
-
-    @Test fun buildPropertiesBundled() {
-        val hexagonProperties = urlOf("classpath:hexagon.properties").readText()
-
-        assertContains(hexagonProperties, "project=")
-        assertContains(hexagonProperties, "module=")
-        assertContains(hexagonProperties, "version=")
-        assertContains(hexagonProperties, "group=")
-        assertContains(hexagonProperties, "description=")
-    }
-
-    @Suppress("RedundantExplicitType") // Type declared for examples generation
-    @Test fun loggerUsage() {
-        // logger
-        val classLogger: Logger = Logger(Runtime::class) // Logger for the `Runtime` class
-        val instanceLogger: Logger = Logger(this::class) // Logger for this instance's class
-
-        logger.info {
-            """
-            You can add a quick log without declaring a Logger with 'com.hexagonkt.helpers.logger'.
-            It is a default logger created for the System class (same as `Logger(System::class)`).
-            """
-        }
-
-        classLogger.trace { "Message only evaluated if trace enabled" }
-        classLogger.debug { "Message only evaluated if debug enabled" }
-        classLogger.warn { "Message only evaluated if warn enabled" }
-        classLogger.info { "Message only evaluated if info enabled" }
-
-        val exception = IllegalStateException("Exception")
-        classLogger.warn(exception) { "Warning with exception" }
-        classLogger.error(exception) { "Error message with exception" }
-        classLogger.warn(exception)
-        classLogger.error(exception)
-        classLogger.error { "Error without an exception" }
-
-        // Logging level can be changed programmatically
-        LoggingManager.setLoggerLevel(ERROR)
-        LoggingManager.setLoggerLevel(classLogger::class, DEBUG)
-        LoggingManager.setLoggerLevel("com.hexagonkt", INFO)
-        // logger
-    }
-}
