--- core/src/test/kotlin/com/hexagonkt/core/text/StringsSamplesTest.kt
@@ -1,13 +1,0 @@
-package com.hexagonkt.core.text
-
-import org.junit.jupiter.api.Test
-
-internal class StringsSamplesTest {
-
-    @Test fun filterVarsExample() {
-        val template = "User {{user}}"
-        val parameters = mapOf<Any, Any>("user" to "John")
-
-        assert (template.filterVars(parameters) == "User John")
-    }
-}
